fix: resolve all 4 pre-merge CI check failures
- fix(ci): correct should_exclude() in code-quality.yml to check
  basename separately so docs/*.md files are properly excluded from
  the 200-line file size check

- fix(migrations): rename 050_remove_triage_from_er.sql to
  063_remove_triage_from_er.sql to eliminate duplicate prefix-050
  conflict with 050_create_ot_procedures_table.sql on main; fixes
  Validate Migration Files and Validate Database Setup checks

- fix(docs): split ADMIN_HANDBOOK.md (was 330 lines) into
  ADMIN_HANDBOOK.md (192 lines) + ADMIN_HANDBOOK_RELEASES.md (106
  lines) to satisfy 200-line limit after the CI bug fix

- fix(docs): trim DATA_FLOW.md from 212 to 191 lines by compressing
  blank lines inside mermaid blocks and fixing the broken/truncated
  section 5

- feat: add docs/ADMIN_HANDBOOK_RELEASES.md, ARCHITECTURE_ENV_CONFIG.md,
  DATA_FLOW_SUPPLEMENTAL.md, DATABASE_SCHEMA_DIAGRAM.md,
  DATABASE_TABLES_1_10.md, DATABASE_SETUP_QUICKSTART.md,
  DATABASE_SETUP_SCHEMA.md

- feat: add src/features/bed-dashboard/lib/triage-wards.ts to keep
  bed-grid-actions.ts within the 200-line limit (192 lines)
